requirements.txt
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
discord==2.3.2
discord.py==2.6.3
dnspython==2.8.0
fonttools==4.60.1
frozenlist==1.7.0
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
multidict==6.6.4
numpy==2.3.3
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pulsefire==2.0.29
pymongo==4.15.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
